+++ playground/src/pages/ScrollControlsDemo.vue
@@ -3,6 +3,8 @@ import { ref, watchEffect } from 'vue'
 import { TresCanvas, useRenderLoop } from '@tresjs/core'
 import { ScrollControls, Stars, Sphere, Box } from '@tresjs/cientos'
 import { SRGBColorSpace, NoToneMapping } from 'three'
+import { TresLeches, useControls } from '@tresjs/leches'
+import '@tresjs/leches/styles'
 
 const scRef = ref()
 const sphereRef = ref()
@@ -20,6 +22,12 @@ const gl = {
   outputColorSpace: SRGBColorSpace,
   toneMapping: NoToneMapping,
 }
+
+useControls('fpsgraph')
+useControls({
+  progress: progress.value,
+})
+
 const { onLoop } = useRenderLoop()
 onLoop(() => {
   if (boxRef.value) {
@@ -28,15 +36,35 @@ onLoop(() => {
   }
 })
 </script>
+
 <template>
-  <TresCanvas v-bind="gl" ref="canvasRef" window-size>
+  <TresLeches class="important-fixed" />
+  <TresCanvas
+    v-bind="gl"
+    window-size
+  >
     <TresPerspectiveCamera :position="[0, 2, 5]" />
     <Stars :radius="1" />
     <TresGridHelper :args="[10, 10]" />
 
-    <ScrollControls ref="scRef" v-model="progress" :distance="10" :smooth-scroll="0.1" html-scroll>
-      <Sphere ref="sphereRef" :scale="0.1" :position="[1, 2, 0]" />
-      <Box ref="boxRef" :scale="0.5" :color="0xff00ff" :position="[-1, 1, 0]" />
+    <ScrollControls
+      ref="scRef"
+      v-model="progress"
+      :distance="10"
+      :smooth-scroll="0.1"
+      html-scroll
+    >
+      <Sphere
+        ref="sphereRef"
+        :scale="0.1"
+        :position="[1, 2, 0]"
+      />
+      <Box
+        ref="boxRef"
+        :scale="0.5"
+        :color="0xff00ff"
+        :position="[-1, 1, 0]"
+      />
     </ScrollControls>
   </TresCanvas>
   <main>
@@ -52,6 +80,7 @@ onLoop(() => {
     </section>
   </main>
 </template>
+
 <style scoped>
 .fixed {
   position: fixed;