+++ playground/src/pages/index.vue
@@ -1,6 +1,7 @@
-<script setup lang="ts"></script>
+<script setup lang="ts"> </script>
+
 <template>
   <Suspense>
-    <OrbitControlsDemo />
+    <GlassMaterialDemo />
   </Suspense>
 </template>